+++ src/main/java/domain/RaceCondition.java
@@ -1,4 +1,4 @@
-package racingcar;
+package domain;
 
 import java.util.Random;
