+++ src/main/java/domain/RaceRecord.java
@@ -1,4 +1,4 @@
-package racingcar;
+package domain;
 
 import java.util.ArrayList;
 import java.util.List;
