+++ src/main/java/domain/RaceWinners.java
@@ -1,4 +1,4 @@
-package racingcar;
+package domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
