+++ src/test/java/domain/RaceConditionTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package domain;
 
 import org.junit.jupiter.api.Test;
